fix stack delete behavior to not attempt to delete configured lambda roles. (Closes #787) (#788)
* Corrections to CONTRIBUTING.md

Running `make test` fails unless the `-modtime` option to `go-bindata` is available.
The jteeuwen fork lacks such an option.
My assumption is that at some point this project migrated from the jteeuwen fork to the shuLhan fork and failed to update the contributing docs.

Also, a minor typo correction.

* Don't attempt to delete a configured lambda role (fixes #787)

This makes the behavior of delete stack match create stack (if there's a role configured, skip doing anything regarding the role).

+++ platform/lambda/lambda.go
@@ -814,6 +814,12 @@ func (p *Platform) roleName() string {
 func (p *Platform) deleteRole(region string) error {
 	name := fmt.Sprintf("%s-function", p.config.Name)
 	c := iam.New(session.New(aws.NewConfig().WithRegion(region)))
+	
+	// role is provided
+	if s := p.config.Lambda.Role; s != "" {
+		log.Debugf("using role from config %s; not deleting", s)
+		return nil
+	}
 
 	_, err := c.DeleteRolePolicy(&iam.DeleteRolePolicyInput{
 		RoleName:   &name,
